+++ src/components/Dialog/Guestbook/index.tsx
@@ -1,4 +1,4 @@
-import { sharedStyles } from "@/styles/shared.styles.ts";
+import { styles } from "@/components/Dialog/Guestbook/styles.ts";
 import * as stylex from "@stylexjs/stylex";
 import { useState } from "react";
 
@@ -16,11 +16,11 @@ export function GuestbookDialog({ onClose }: GuestbookDialogProps) {
   };
 
   return (
-    <div {...stylex.props(sharedStyles.overlay)}>
+    <div {...stylex.props(styles.overlay)}>
       <div {...stylex.props(styles.dialog)}>
-        <div {...stylex.props(sharedStyles.header)}>
-          <h2 {...stylex.props(sharedStyles.title)}>
-            방명록 <span {...stylex.props(sharedStyles.titleNum)}>5</span>
+        <div {...stylex.props(styles.header)}>
+          <h2 {...stylex.props(styles.title)}>
+            방명록 <span {...stylex.props(styles.titleNum)}>5</span>
           </h2>
           <img
             onClick={onClose}
@@ -36,14 +36,14 @@ export function GuestbookDialog({ onClose }: GuestbookDialogProps) {
             placeholder="별명"
             value={name}
             onChange={(e) => setName(e.target.value)}
-            {...stylex.props(sharedStyles.input)}
+            {...stylex.props(styles.input)}
           />
           <input
             type="text"
             placeholder="댓글 작성하기"
             value={message}
             onChange={(e) => setMessage(e.target.value)}
-            {...stylex.props(sharedStyles.input)}
+            {...stylex.props(styles.input)}
           />
           <button type="submit" {...stylex.props(styles.submitButton)}>
             작성
@@ -69,76 +69,3 @@ export function GuestbookDialog({ onClose }: GuestbookDialogProps) {
     </div>
   );
 }
-
-export const styles = stylex.create({
-  dialog: {
-    width: "100%",
-    maxWidth: "430px",
-    backgroundColor: "#fff",
-    borderRadius: "16px 16px 0 0",
-    position: "fixed",
-    bottom: 0,
-    padding: "20px",
-    maxHeight: "90vh",
-    overflowY: "auto",
-  },
-  formContainer: {
-    display: "flex",
-    gap: "8px",
-    flexDirection: "column",
-    marginBottom: "16px",
-  },
-  submitButton: {
-    width: "100%",
-    padding: "14px",
-    backgroundColor: "#FF731D",
-    color: "#fff",
-    border: "none",
-    borderRadius: "8px",
-    fontSize: "16px",
-    fontWeight: "600",
-    cursor: "pointer",
-    marginBottom: "20px",
-  },
-  comments: {
-    display: "flex",
-    flexDirection: "column",
-    padding: "16px 0px",
-    gap: "16px",
-  },
-  commentItem: {
-    borderBottom: "1px solid #DDE1E6",
-    paddingBottom: "16px",
-  },
-  commentHeader: {
-    display: "flex",
-    justifyContent: "space-between",
-    alignItems: "center",
-    marginBottom: "8px",
-  },
-  editButton: {
-    color: "#FF731D",
-    background: "none",
-    border: "none",
-    padding: 0,
-    fontSize: "14px",
-    fontWeight: 400,
-    cursor: "pointer",
-  },
-  date: {
-    fontSize: "12px",
-    fontWeight: 400,
-    color: "#909090",
-  },
-  commentText: {
-    fontSize: "14px",
-    lineHeight: "1.5",
-    color: "#333",
-    whiteSpace: "pre-wrap",
-  },
-  closeIcon: {
-    width: 24,
-    height: 24,
-    cursor: "pointer",
-  },
-});
