+++ src/components/Dialog/Password/index.tsx
@@ -1,11 +1,7 @@
-import { sharedStyles } from "@/styles/shared.styles.ts";
+import { styles } from "@/components/Dialog/Password/styles.ts";
+import type { PasswordDialogProps } from "@/components/Dialog/Password/types.ts";
 import stylex from "@stylexjs/stylex";
 import { useState } from "react";
-
-interface PasswordDialogProps {
-  onClose: () => void;
-  onConfirm: (password: string) => void;
-}
 
 export function PasswordDialog({ onClose, onConfirm }: PasswordDialogProps) {
   const [password, setPassword] = useState("");
@@ -16,7 +12,7 @@ export function PasswordDialog({ onClose, onConfirm }: PasswordDialogProps) {
   };
 
   return (
-    <div {...stylex.props(sharedStyles.overlay)}>
+    <div {...stylex.props(styles.overlay)}>
       <div {...stylex.props(styles.dialog)}>
         <div {...stylex.props(styles.header)}>
           <img
@@ -34,7 +30,7 @@ export function PasswordDialog({ onClose, onConfirm }: PasswordDialogProps) {
             placeholder="비밀번호를 입력하세요"
             value={password}
             onChange={(e) => setPassword(e.target.value)}
-            {...stylex.props(sharedStyles.input)}
+            {...stylex.props(styles.input)}
           />
           <button type="submit" {...stylex.props(styles.confirmButton)}>
             확인
@@ -44,42 +40,3 @@ export function PasswordDialog({ onClose, onConfirm }: PasswordDialogProps) {
     </div>
   );
 }
-
-export const styles = stylex.create({
-  dialog: {
-    width: "60%",
-    backgroundColor: "#fff",
-    borderRadius: "16px",
-    padding: "20px",
-  },
-  title: {
-    textAlign: "center",
-    fontSize: "18px",
-    fontWeight: 600,
-    paddingBottom: "20px",
-  },
-  formContainer: {
-    display: "flex",
-    flexDirection: "column",
-    gap: "12px",
-  },
-  header: {
-    display: "flex",
-    justifyContent: "flex-end",
-    marginBottom: "4px",
-  },
-  confirmButton: {
-    width: "100%",
-    padding: "14px",
-    backgroundColor: "#f1f3f5",
-    border: "none",
-    borderRadius: "8px",
-    fontSize: "16px",
-    cursor: "pointer",
-  },
-  closeIcon: {
-    width: 24,
-    height: 24,
-    cursor: "pointer",
-  },
-});
