+++ src/components/Dialog/Password/types.ts
@@ -1,0 +1,4 @@
+export interface PasswordDialogProps {
+  onClose: () => void;
+  onConfirm: (password: string) => void;
+}
