+++ src/components/ImageUpload/types.ts
@@ -1,0 +1,7 @@
+import type { UseFormRegisterReturn } from "react-hook-form";
+
+export interface ImageUploadProps {
+  register: UseFormRegisterReturn;
+  onClearError: () => void;
+  error?: string;
+}
